afb-hsrv: Prepare selection of listening interfaces
The listening socket will no more be the default one
but will be the selected ones, with ability to select
more than just one listening interface.

Nevertheless, the behavior is not changed at the moment.
The default interface (0.0.0.0) is still used. A further
change will add the ability to select the listening
interfaces.

That evolution depends on a fix of libmicrohttpd to
enable at the same time epoll and no-listen-fd.

Bug-AGL: SPEC-1833
Bug-AGL: SPEC-1016

Change-Id: Ia78671cbee90a243ba7b2c724b6155cffbde6920
Signed-off-by: José Bollo <jose.bollo@iot.bzh>