afb-daemon: Add /tmp has fallback for uploads
During its start, when implementing HTTP server,
afb-daemon checked that the upload directory was
accessible and writable and afb-daemon refused to
start when it was not the case.

When trying to use systemd's DynamicUser for platform
services, it had the effect to forbid service to run.

That commit add the directory /tmp as a fallback
upload directory.

Bug-AGL: SPEC-2446

Signed-off-by: José Bollo <jose.bollo@iot.bzh>
Change-Id: Ib4ee456d71cb1aefa19e8eed6774ef5f77919366