afb-client-demo: Allow pipe of messages
The tool afb-client-demo implements a synchronous
mode (option -s) that allows to wait for a reply
before to send the next request.

It is interesting for the purpose of testing to allow
afb-client-demo it to send a controlled count of
requests, ensuring that there is a controlled count
of pending requests.

Bug-AGL: SPEC-2638

Signed-off-by: Jose Bollo <jose.bollo@iot.bzh>
Change-Id: Ic854d68aaeb82c17b4640c5c9de46181ad3a7e5e