afb-proto-ws: Change the protocol WSAPI
Change internals of the protocol WSAPI for the following
rationale:

 1. Enforce specific declaration and transmission of session
    identifiers and of access tokens.
 2. Lower the size of identifiers to be 16 bits.
 3. Introduce protocol versionning through a mechanism of offer/set.

The main purpose of that change is to optimize the count of data
transmitted. It manages as best as possible the transmission of
access tokens the less possible times. Same for sessions that
the chage was transmitted at each call.

Bug-AGL: SPEC-2968

Change-Id: If0a22b86627ead35a410e51c1028025c5b02c38f
Signed-off-by: Jose Bollo <jose.bollo@iot.bzh>