afb-hswitch: Update websocket policy
The previous policy for websocket was to enforce them
to present a valid token when establishing. This policy
is removed because the Websocket API of javascript doesn't
provide a way to set the Bearer token in the HTTP header
of the negociation (a big miss in my opinion).

But because the new policy of token logic is to check
the token at each request, it is not more needed to check
it at establishment.

Bug-AGL: SPEC-2968

Change-Id: I2941757492a27a2eed14e26fbb411330ab4aa8bc
Signed-off-by: Jose Bollo <jose.bollo@iot.bzh>