afb-context: Move credentials to context
The split between context and credentials
in requests was somehow artificial and awkward.

This change move the credentials to the context
and removes as many references to credentials
as possible in favor of working on contexts.

Change the value returned by afb_auth_check to be 1
if validated, 0 or less than zero if not validated.

Bug-AGL: SPEC-2968

Signed-off-by: José Bollo <jose.bollo@iot.bzh>
Change-Id: I979dc841e03247e126e3fa8433a1cc0d4108adf0