afb-daemon: Fix optionnal port
Since:
 - introduction of interface to set the
   listening port(s) and address(es)
 - use of nss-localuser for discriminating
   binders

there is no more strict need for exporting
the port and substituting it. So the
substitution is make only if needed.

Since removal of token generation, there is no
more need to propoagate any token.

The two possibilities are kept for being used
whene needed and to ensure sweet transition.

Bug-AGL: SPEC-2968

Signed-off-by: José Bollo <jose.bollo@iot.bzh>
Change-Id: I1d2b4d765f244e46de3071baddd8b3d0817b9436