afb-evt: Ensure unsubscribe works
Fix the logic of unsubscribing to events. It
was not possible before to implment it without
tracking every session and context. It was not
done because of the required complexity.

This implementation ensures that unexpected
events lead to a removal of the listener from the
list of watchers of the events.

The management of the list of watchers is reworked
to free unused memory.

Bug-AGL: SPEC-3069

Change-Id: Ie67372adbde9dcb9dc6c5c2738111d22609e7256
Signed-off-by: Jose Bollo <jose.bollo@iot.bzh>